azure-enterprise-platform/requirements.txt
# Add your Python dependencies here
azure-identity
azure-keyvault-secrets
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
azure-mgmt-storage